strip v2 prefix
